refactor(test): move inside `remote/`
